Move checkpoint saving util
