+++ tests/test_base_component_loading.py
@@ -5,7 +5,7 @@ import torch
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
 from components.hierarchical_autoencoder import HierarchicalAutoencoder
-from train import save_base_components
+from components.checkpoint_utils import save_base_components
 
 
 def build_base_model():
